+++ packages/giselle/src/engine/utils/workflow/slice-graph-from-node.ts
