+++ packages/giselle/src/engine/utils/workflow/test/fixtures/workspace1.json
