+++ packages/giselle/src/engine/utils/workflow/test/fixtures/workspace2.json
