+++ packages/giselle/src/engine/utils/workflow/test/fixtures/workspace3.json
