+++ packages/giselle/src/engine/utils/workflow/test/fixtures/workspace4.json
