--- packages/workflow-utils/package.json
@@ -1,32 +1,0 @@
-{
-	"name": "@giselle-sdk/workflow-utils",
-	"version": "0.0.0",
-	"private": true,
-	"main": "./dist/index.js",
-	"module": "./dist/index.mjs",
-	"types": "./dist/index.d.ts",
-	"scripts": {
-		"build": "tsup",
-		"check-types": "tsc --noEmit",
-		"clean": "rm -rf .turbo dist node_modules",
-		"format": "biome check --write .",
-		"test": "vitest run"
-	},
-	"exports": {
-		"./package.json": "./package.json",
-		".": {
-			"types": "./dist/index.d.ts",
-			"import": "./dist/index.mjs",
-			"require": "./dist/index.js"
-		},
-		"./internal": "./src/index.ts"
-	},
-	"devDependencies": {
-		"@giselle/giselle-sdk-tsconfig": "workspace:*",
-		"tsup": "catalog:",
-		"vitest": "catalog:"
-	},
-	"dependencies": {
-		"@giselle-sdk/data-type": "workspace:*"
-	}
-}
