--- packages/workflow-utils/src/index.ts
@@ -1,1 +1,0 @@
-export * from "./build-workflow-from-node";
