--- packages/workflow-utils/tsconfig.json
@@ -1,8 +1,0 @@
-{
-	"extends": "@giselle/giselle-sdk-tsconfig/react-library.json",
-	"include": ["."],
-	"exclude": ["node_modules"],
-	"compilerOptions": {
-		"lib": ["ESNext"]
-	}
-}
